chore(logs): move all filter components into a single directory and update import paths
