+++ products/logs/frontend/components/LogsViewer/Filters/LogsFilterBar.tsx
@@ -9,15 +9,15 @@ import { useFeatureFlag } from 'lib/hooks/useFeatureFlag'
 import { IconPauseCircle, IconPlayCircle } from 'lib/lemon-ui/icons'
 import { Scene } from 'scenes/sceneTypes'
 
-import { LogsFilterGroup } from 'products/logs/frontend/components/filters/LogsFilters/FilterGroup'
-import { DateRangeFilter } from 'products/logs/frontend/filters/DateRangeFilter'
-import { FilterHistoryDropdown } from 'products/logs/frontend/filters/FilterHistoryDropdown'
-import { LogsDateRangePicker } from 'products/logs/frontend/filters/LogsDateRangePicker/LogsDateRangePicker'
-import { ServiceFilter } from 'products/logs/frontend/filters/ServiceFilter'
-import { SeverityLevelsFilter } from 'products/logs/frontend/filters/SeverityLevelsFilter'
+import { DateRangeFilter } from 'products/logs/frontend/components/LogsViewer/Filters/DateRangeFilter'
+import { LogsFilterGroup } from 'products/logs/frontend/components/LogsViewer/Filters/FilterGroup'
+import { FilterHistoryDropdown } from 'products/logs/frontend/components/LogsViewer/Filters/FilterHistoryDropdown'
+import { LogsDateRangePicker } from 'products/logs/frontend/components/LogsViewer/Filters/LogsDateRangePicker/LogsDateRangePicker'
+import { ServiceFilter } from 'products/logs/frontend/components/LogsViewer/Filters/ServiceFilter'
+import { SeverityLevelsFilter } from 'products/logs/frontend/components/LogsViewer/Filters/SeverityLevelsFilter'
 import { logsLogic } from 'products/logs/frontend/logsLogic'
 
-export const LogsFilters = (): JSX.Element => {
+export const LogsFilterBar = (): JSX.Element => {
     const newLogsDateRangePicker = useFeatureFlag('NEW_LOGS_DATE_RANGE_PICKER')
     const { logsLoading, liveTailRunning, liveTailDisabledReason, dateRange } = useValues(logsLogic)
     const { runQuery, zoomDateRange, setLiveTailRunning, setDateRange } = useActions(logsLogic)
