+++ products/logs/frontend/components/LogsViewer/Filters/zoom-utils.ts
